fix: split transaction-service into core-wallet-services package
