+++ core/wallet-services/package.json
@@ -24,10 +24,20 @@
         "test": "vitest run --project node --project browser --passWithNoTests",
         "test:coverage": "vitest run --project node --project browser --coverage --passWithNoTests"
     },
-    "dependencies": {},
+    "dependencies": {
+        "@canton-network/core-signing-lib": "workspace:^",
+        "@canton-network/core-signing-securosys": "workspace:^",
+        "@canton-network/core-wallet-auth": "workspace:^",
+        "@canton-network/core-wallet-store": "workspace:^",
+        "@canton-network/core-wallet-user-rpc-client": "workspace:^",
+        "@logtape/logtape": "^2.3.1",
+        "pino": "^10.3.1",
+        "uuid": "^14.0.2"
+    },
     "devDependencies": {
         "@vitest/browser-playwright": "^4.1.10",
         "@vitest/coverage-v8": "^4.1.10",
+        "pino-test": "^1.1.0",
         "playwright": "^1.61.1",
         "tsdown": "^0.22.9",
         "typescript": "^5.9.3",
