+++ core/wallet-services/src/index.ts
@@ -1,8 +1,6 @@
 // Copyright (c) 2025-2026 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
+export * from './transaction-service'
 export * from './types'
-export * from './errors'
-export * from './events'
-export { DiscoveryClient } from './client'
-export type { DiscoveryClientConfig, ActiveSession } from './client'
+export * from './utils'
