+++ core/wallet-services/src/types.ts
@@ -1,87 +1,27 @@
 // Copyright (c) 2025-2026 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-import type { Provider } from '@canton-network/core-splice-provider'
-import type {
-    WalletPickerEntry,
-    WalletPickerResult,
-} from '@canton-network/core-types'
-export type {
-    WalletPickerEntry,
-    WalletPickerResult,
-} from '@canton-network/core-types'
-import type {
-    ProviderId,
-    ProviderType,
-    RpcTypes as DappRpcTypes,
-} from '@canton-network/core-wallet-dapp-rpc-client'
-
-export interface WalletInfo {
-    providerId: ProviderId
-    name: string
-    type: ProviderType
-    description?: string | undefined
-    icon?: string | undefined
-    url?: string | undefined
-    /**
-     * If set, wallet picker keeps the global popup open after pick for reuse
-     * by async navigations (e.g. HTTP wallet gateway). Not used for sync DApp API
-     * wallets even when `type` is `remote`.
-     */
-    reuseGlobalWalletPopup?: boolean | undefined
-}
+import {
+    SigningDriverInterface,
+    SigningProvider,
+} from '@canton-network/core-signing-lib'
 
 /**
- * A ProviderAdapter is a thin factory that knows how to create a
- * Provider<DappRpcTypes> for a particular wallet type, detect its
- * availability, and clean up resources.
- *
- * All RPC methods (connect, disconnect, status, prepareExecute, etc.)
- * are called directly on the provider — the adapter does not duplicate them.
+ * Signing drivers registered at Wallet Gateway startup.
+ * Partial because only configured providers are present (e.g. Dfns when env vars are set).
  */
-export interface ProviderAdapter {
-    readonly providerId: ProviderId
-    readonly name: string
-    readonly type: ProviderType
-    readonly icon?: string | undefined
+export type SigningDrivers = Partial<
+    Record<SigningProvider, SigningDriverInterface>
+>
 
-    /** Return wallet metadata for display in the wallet picker. */
-    getInfo(): WalletInfo
+type EventListener = (...args: unknown[]) => void
+export interface Notifier {
+    on(event: string, listener: EventListener): void
 
-    /**
-     * Check whether this wallet is currently available.
-     * Extensions probe for the browser extension; gateways always return true.
-     */
-    detect(): Promise<boolean>
+    emit(event: string, ...args: unknown[]): boolean
 
-    /**
-     * Return a Provider<DappRpcTypes> for this wallet.
-     * Extension adapters return a DappProvider (openrpc-dapp-api.json).
-     * Remote adapters may return a provider that bridges the
-     * remote API to the dApp API surface.
-     *
-     * The caller is responsible for invoking `provider.request({ method: 'connect' })`
-     * and `provider.request({ method: 'disconnect' })`.
-     */
-    provider(): Provider<DappRpcTypes>
-
-    /**
-     * Clean up adapter-specific resources (e.g. close popup windows).
-     * Called after disconnect. Does NOT call disconnect on the provider.
-     */
-    teardown(): void
-
-    /**
-     * Attempt to restore a previous session (e.g. from localStorage).
-     * Returns a ready-to-use provider if the session was restored, or null.
-     */
-    restore?(): Promise<Provider<DappRpcTypes> | null>
+    removeListener(event: string, listenerToRemove: EventListener): void
 }
 
-/**
- * A function that presents wallet choices to the user and returns their selection.
- * This abstraction allows the DiscoveryClient to remain UI-framework agnostic.
- */
-export type WalletPickerFn = (
-    entries: WalletPickerEntry[]
-) => Promise<WalletPickerResult>
+export type HASHING_SCHEME_VERSION =
+    'HASHING_SCHEME_VERSION_V2' | 'HASHING_SCHEME_VERSION_V3'
