+++ wallet-gateway/remote/package.json
@@ -53,6 +53,7 @@
         "@canton-network/core-types": "workspace:^",
         "@canton-network/core-wallet-auth": "workspace:^",
         "@canton-network/core-wallet-dapp-rpc-client": "workspace:^",
+        "@canton-network/core-wallet-services": "workspace:*",
         "@canton-network/core-wallet-store": "workspace:^",
         "@canton-network/core-wallet-store-sql": "workspace:*",
         "@canton-network/core-wallet-ui-components": "workspace:^",
